+++ src/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import { useRoute } from 'vue-router';
+import { useRoute, useRouter } from 'vue-router';
 import Toast from 'primevue/toast';
+import { App } from '@capacitor/app';
 
 import { useDatabaseStore } from '@/stores/database';
 import { useMapsStore } from '@/stores/maps';
@@ -9,12 +10,18 @@ import FirstStart from '@/components/FirstStart.vue';
 import MainToolbar from '@/components/MainToolbar.vue';
 
 const route = useRoute();
+const router = useRouter();
 const databaseStore = useDatabaseStore();
 const mapsStore = useMapsStore();
 
 (async () => {
   const res = await db.vehicles.count();
   databaseStore.setEmpty(!res);
+
+  App.addListener('backButton', async ({ canGoBack }) => {
+    if (canGoBack) router.back();
+    else await App.exitApp();
+  });
 })();
 </script>
 
@@ -41,10 +48,12 @@ const mapsStore = useMapsStore();
 .main-container {
   background-color: white;
 }
+
 .main-content {
   height: 87vh;
   overflow-y: scroll;
 }
+
 .transparent {
   background-color: transparent !important;
 }